requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
python-dotenv
httpx
scipy